the same-size slice, all 69 classified, and three of them were real
`compare.py` counts the "same size, different instructions" bucket and never
printed it, so the sharpest slice in the object -- same byte count, so nothing
missing and nothing extra -- had never been enumerated.  `tools/samesize.py`
prints it, `--all` dumps every aligned diff in one pass, and finding 2900
classifies all 69 rows (56 distinct bodies) as forced, free or in between.

Three were forced and are fixed, both provably invisible to every test and so
codegen-tier-only by construction:

  alaw2ulaw, ulaw2alaw   a `& 0x7f` the original does not have.  Each arm's
                         xor clears bit 7 already, so the mask is redundant
                         over all 256 inputs; `and $0x7f` against the blob's
                         `movzbl %al` is the only difference
  dtmf_progress          two locals declared `short` that the object declares
                         `int` -- a 32-bit compare and a `movswl` on return

  334 -> 337 identical, set-diffed on both sides: +3, -0, nothing lost.

The more useful half is what was declined.  `resetLinearMapping` differs by
exactly `movzwl` against `movswl` and looks like 613 -- and the consumer is a
16-bit store, so it is 614, and retyping `ucodeLevel` would have flipped every
negative linear level.  Findings 2902 and 2903 record that, the nine-row
sibling-call group (same callee both sides, no hypothesis), and eight more
forced rows named with the instruction that shows each.

Two of the session's own errors are in the record rather than edited out: a
`.rodata` offset read as a string LENGTH when it was the arm ORDER, and a
class table that summed to 69 while double-counting three rows.

`make phase`: 5 ALLOWED, 10 FAIL (the declared gccdiverge entries), period
differential 185 passed, 0 failed.

+++ src/service/dtmf.c
@@ -320,11 +320,29 @@ dtmf_detect(float x, struct dtmf *d, short mode)
 short
 dtmf_progress(struct dtmf *d, const float *samples, short count, short mode)
 {
-	short result = DTMF_NO_DIGIT;
+	/*
+	 * BOTH LOCALS ARE `int` AND THE OBJECT IS WHAT SAYS SO.  The `(short)`
+	 * narrowing is on the CALL, not on the variable, and the two readings
+	 * agree over every value `dtmf_detect` can return -- so no test can
+	 * separate them and the codegen tier had to.  Two instructions carry
+	 * it, and each is a byte's worth of the same claim:
+	 *
+	 *   blob:  cmp $0x1,%ecx           `digit` is int, so the promoted
+	 *   ours:  cmp $0x1,%cx            compare is 32-bit, not 16
+	 *
+	 *   blob:  movswl 0x18(%esp),%eax  `result` is int and the RETURN
+	 *   ours:  mov    0x18(%esp),%eax  type is short, so returning it
+	 *                                  truncates; a short `result` needs
+	 *                                  no re-extension and gets none
+	 *
+	 * The 0x66 prefix on the narrow compare pays for the byte the narrow
+	 * load saves, which is how this stayed the right size.  Finding 2901.
+	 */
+	int result = DTMF_NO_DIGIT;
 	int i;
 
 	for (i = 0; i < count; i++) {
-		short digit = (short)dtmf_detect(samples[i], d, mode);
+		int digit = (short)dtmf_detect(samples[i], d, mode);
 
 		if (digit != DTMF_NOT_YET && digit != DTMF_NO_DIGIT)
 			result = digit;
